Fix compilation warnings

+++ rplidar/sdk/src/rplidar_driver.cpp
@@ -1205,7 +1205,7 @@ void RPlidarDriverImplCommon::_ultraCapsuleToNormal(const rplidar_response_ultra
 
            
             dist_q2[0] = (dist_major << 2);
-            if ((dist_predict1 == 0xFFFFFE00) || (dist_predict1 == 0x1FF)) {
+            if (((unsigned int)dist_predict1 == 0xFFFFFE00) || ((unsigned int)dist_predict1 == 0x1FF)) {
                 dist_q2[1] = 0;
             } else {
                 dist_predict1 = (dist_predict1 << scalelvl1);
@@ -1213,7 +1213,7 @@ void RPlidarDriverImplCommon::_ultraCapsuleToNormal(const rplidar_response_ultra
 
             }
 
-            if ((dist_predict2 == 0xFFFFFE00) || (dist_predict2 == 0x1FF)) {
+            if (((unsigned int)dist_predict2 == 0xFFFFFE00) || ((unsigned int)dist_predict2 == 0x1FF)) {
                 dist_q2[2] = 0;
             } else {
                 dist_predict2 = (dist_predict2 << scalelvl2);
